feat: Handle integer overflow by casting multiplication to long and make test pass.

+++ JavaDsaWithTest/src/main/java/org/practice/dsa/leet_code/easy/QuestionSquareRoot.java
@@ -9,9 +9,9 @@ public class QuestionSquareRoot {
         while (start <= end) {
             int mid = start + (end - start) / 2;
             
-            if (mid * mid == x) {
+            if (((long) mid * mid) == x) {
                 return mid;
-            } else if (mid * mid <= x) {
+            } else if (((long)mid * mid) <= x) {
                 start = mid + 1;
             } else {
                 end = mid - 1;
